+++ example/android/app/src/main/kotlin/com/example/mem_value_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.mem_value_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()